Clear saved text in the dom when clicking Clear button
Prior to this change, when a user clicked the Clear saved text button we cleared the clippings list. However, the DOM didn't reflext the change and continued to display the list of clippings. This adds a line to update the DOM as well as send the clear message to chrome.runtime. This way, the user gets immediate feedback for their action.