Add the X button to the UI and create the delete handler on the FE
This change adds the delete button to each clip item in the UI. It creates an event listener on the button which on click calls the onDeleteClick function, passing it the creation date which is used as a unique id. The onDeleteClick function then sends the message passing the id so it can be removed from the clippings array.